+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/api/pushrules/EventMatchCondition.kt
@@ -18,6 +18,9 @@ package org.matrix.android.sdk.api.pushrules
 
 import org.matrix.android.sdk.api.session.events.model.Event
 import org.matrix.android.sdk.internal.di.MoshiProvider
+import org.matrix.android.sdk.internal.util.caseInsensitiveFind
+import org.matrix.android.sdk.internal.util.hasSpecialGlobChar
+import org.matrix.android.sdk.internal.util.simpleGlobToRegExp
 import timber.log.Timber
 
 class EventMatchCondition(
@@ -29,16 +32,18 @@ class EventMatchCondition(
          * The glob-style pattern to match against. Patterns with no special glob characters should
          * be treated as having asterisks prepended and appended when testing the condition.
          */
-        val pattern: String
-) : Condition(Kind.EventMatch) {
+        val pattern: String,
+        /**
+         * true to match only words. In this case pattern will not be considered as a glob
+         */
+        val wordsOnly: Boolean
+) : Condition {
 
     override fun isSatisfied(event: Event, conditionResolver: ConditionResolver): Boolean {
         return conditionResolver.resolveEventMatchCondition(event, this)
     }
 
-    override fun technicalDescription(): String {
-        return "'$key' Matches '$pattern'"
-    }
+    override fun technicalDescription() = "'$key' matches '$pattern', words only '$wordsOnly'"
 
     fun isSatisfied(event: Event): Boolean {
         // TODO encrypted events?
@@ -48,14 +53,18 @@ class EventMatchCondition(
 
         // Patterns with no special glob characters should be treated as having asterisks prepended
         // and appended when testing the condition.
-        try {
-            val modPattern = if (hasSpecialGlobChar(pattern)) simpleGlobToRegExp(pattern) else simpleGlobToRegExp("*$pattern*")
-            val regex = Regex(modPattern, RegexOption.DOT_MATCHES_ALL)
-            return regex.containsMatchIn(value)
+        return try {
+            if (wordsOnly) {
+                value.caseInsensitiveFind(pattern)
+            } else {
+                val modPattern = if (pattern.hasSpecialGlobChar()) pattern.simpleGlobToRegExp() else "*$pattern*".simpleGlobToRegExp()
+                val regex = Regex(modPattern, RegexOption.DOT_MATCHES_ALL)
+                regex.containsMatchIn(value)
+            }
         } catch (e: Throwable) {
             // e.g PatternSyntaxException
             Timber.e(e, "Failed to evaluate push condition")
-            return false
+            false
         }
     }
 
@@ -77,28 +86,5 @@ class EventMatchCondition(
             }
         }
         return null
-    }
-
-    companion object {
-
-        private fun hasSpecialGlobChar(glob: String): Boolean {
-            return glob.contains("*") || glob.contains("?")
-        }
-
-        // Very simple glob to regexp converter
-        private fun simpleGlobToRegExp(glob: String): String {
-            var out = "" // "^"
-            for (element in glob) {
-                when (element) {
-                    '*'  -> out += ".*"
-                    '?'  -> out += '.'.toString()
-                    '.'  -> out += "\\."
-                    '\\' -> out += "\\\\"
-                    else -> out += element
-                }
-            }
-            out += "" // '$'.toString()
-            return out
-        }
     }
 }